Fix narration value of reconciled entries
Changes how narration field is defined on reconciled  entires so
that it now agrees with documentation. Now narration value
is always set  as defined on the expected txn (previously this was
only the case if the narration was not otherwise defined on the
incoming txn). If narration is not defined on the expected txn
then any narration defined on the incoming txn will be
preserved.